move `OverWindowExecutor` to `eowc` module
